Refactor Team workflow code as it became big and hard to maintain
